Update User Validator

+++ CatalogueApp/lab11/src/main/java/OnlineCatalogue/Controller/UserController.java
@@ -3,8 +3,11 @@ package OnlineCatalogue.Controller;
 import OnlineCatalogue.Mapper.Mapper;
 import OnlineCatalogue.DTOs.UserDTOs.CreateUserDTO;
 import OnlineCatalogue.DTOs.UserDTOs.UserDTO;
+import OnlineCatalogue.Response.ApiResponse;
 import OnlineCatalogue.Service.UserService;
+import OnlineCatalogue.Validator.UserValidator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -25,8 +28,12 @@ public class UserController {
     }
 
     @GetMapping("/get/user")
-    public UserDTO GetUser(@RequestParam(name = "id") int id){
-        return mapper.ToUserDTO(userService.FindById(id));
+    public ApiResponse<UserDTO> GetUser(@RequestParam(name = "id") int id){
+        UserValidator userValidator = new UserValidator();
+        if(userValidator.ValidateId(id).equals(ResponseEntity.ok("Id found."))){
+            return new ApiResponse<UserDTO>("User found", mapper.ToUserDTO(userService.FindById(id)));
+        }
+        return new ApiResponse<UserDTO>("User not found.");
     }
 
     @PostMapping("/add/user")
@@ -38,15 +45,25 @@ public class UserController {
     }
 
     @PutMapping("/update/user")
-    public UserDTO UpdateUser(@RequestParam(name = "id") int id,
-                              @RequestParam(name = "username") String username,
-                              @RequestParam(name = "password") String password){
-        return mapper.ToUserDTO(userService.UpdateUser(id, new CreateUserDTO(username,password)));
+    public ResponseEntity<?> UpdateUser(@RequestParam(name = "id") int id,
+                                             @RequestParam(name = "username") String username,
+                                             @RequestParam(name = "password") String password){
+        UserValidator userValidator = new UserValidator();
+        if(userValidator.ValidateId(id).equals(ResponseEntity.ok("Id found."))) {
+           userService.UpdateUser(id, new CreateUserDTO(username, password));
+           return ResponseEntity.ok().body("User updated successfully.");
+        }
+        return ResponseEntity.badRequest().body("There is no user with id " + id);
     }
 
     @DeleteMapping("/delete/user")
-    public boolean DeleteUser(@RequestParam(name = "id") int id){
-        userService.DeleteUSer(id);
-        return true;
+    public ResponseEntity<?> DeleteUser(@RequestParam(name = "id") int id){
+        UserValidator userValidator = new UserValidator();
+        ResponseEntity<?> userResponse = userValidator.ValidateId(id);
+        if(userResponse.equals(ResponseEntity.ok("Id found."))) {
+            userService.DeleteUSer(id);
+            return ResponseEntity.ok("User deleted successfully");
+        }
+        return userResponse;
     }
 }
